SymmetricTree.js (#69)
In Binary Trees Data Structure, I've proposed a solution to one of the most renowned interview questions asked in the "FAANG" firms. I employed the "Depth First Search" traversal technique to check if it is a Symmetric Tree. Given the root of a binary tree, we check whether it is a mirror of itself (i.e., symmetric around its center).