SameTree.js
In Binary Trees Data Structure, I've proposed a solution to one of the most renowned interview questions asked in the "FAANG" firms. I employed the "Depth First Search" traversal technique to check if they are the same tree. Given the roots of two binary trees p and q, write a function to check if they are the same or not. Two binary trees are considered the same if they are structurally identical, and the nodes have the same value.